Reword the section on what HTTPS does not protect
I'm not super happy with this rewording, since it's not absolutely accurate, yet also wordier. But it bothered me a bit that the section and the previous section imply that the HTTPS connection does not encrypt the hostname. It encrypts the hostname in the Host header or the fully qualified URI (if provided instead of a relative URI), since this is all part of the HTTP request itself. The domain/subdomain leakage comes from the DNS lookup and SNI and possibly the server certificate, depending on what names are specified.

I don't have a good solution here though. I think my suggested wording is marginally better, but I don't really like it either, and would be fine with it getting rejected.