Deprecate TLSv1.0 and reference NIST SP 800-52Rev1
If this is intended to be a reference for anyone implementing a web service for the US federal government, then we have to make reference to NIST SP 800-52Rev1.  If we don't want to quote from this document, then we should at least make a reference to it towards the top of our documentation, and then make sure that we follow all the appropriate recommendations.

Moreover, we should warn people about the holes in TLSv1.0 and push them to finally make the move to TLS 1.2.

There's lots more to come.